优惠券领取成功

+++ app/Repositories/Promotion/Rule/EloquentRuleRepository.php
@@ -7,27 +7,29 @@ use App\Models\Promotion\RuleQualify;
 class EloquentRuleRepository implements RuleRepositoryContract {
 
 
-    public function createRule($qualifies, $items, $range, $discount, $weight, $multi, $memo)
+    public function createRule($name, $desc, $qualifies, $items, $range, $discount, $weight, $multi)
     {
-        return $rule = $this->updateOrCreate(null, $qualifies, $items, $range, $discount, $weight, $multi, $memo);
+        return $rule = $this->updateOrCreate(null, $name, $desc, $qualifies, $items, $range, $discount, $weight, $multi);
     }
 
-    protected function updateOrCreate($rule_id = null, $qualifies, $items, $range, $discount, $weight, $multi, $memo)
+    protected function updateOrCreate($rule_id = null, $name, $desc, $qualifies, $items, $range, $discount, $weight, $multi)
     {
         $rule_data = [
+            'name' => $name,
+            'desc' => $desc,
             'qua_type' => $qualifies['type'],
             'qua_quantity' => array_get($qualifies, 'quantity', 1),
-            'qua_content' => $qualifies['values'],
+            'qua_content' => json_encode($qualifies['values']),
             'item_type' => $items['type'],
+            'item_content' => json_encode($items['values']),
             'range_type' => $range['type'],
             'range_min' => $range['min'],
             'range_max' => array_get($range, 'max', null),
             'discount_resource' => $discount['type'],
             'discount_mode' => $discount['mode'],
-            'discount_content' => $discount['value'],
+            'discount_content' => json_encode($discount['value']),
             'weight' => $weight,
-            'multi_type' => $multi,
-            'memo' => $memo
+            'multi_able' => $multi,
         ];
         $rule = is_null($rule_id) ? new Rule() : $this->getRule($rule_id);
         $rule->fill($rule_data);
@@ -37,14 +39,18 @@ class EloquentRuleRepository implements RuleRepositoryContract {
         return $rule;
     }
 
+    /**
+     * @param $rule_id
+     * @return Rule
+     */
     public function getRule($rule_id)
     {
-        return $rule_id instanceof Rule ? $rule_id : Rule::find($rule_id);
+        return $rule_id instanceof Rule ? $rule_id : Rule::query()->find($rule_id);
     }
 
-    public function updateRule($rule_id, $qualifies, $items, $range, $discount, $weight, $multi, $memo)
+    public function updateRule($rule_id, $name, $desc, $qualifies, $items, $range, $discount, $weight, $multi)
     {
-        return $this->updateOrCreate($rule_id, $qualifies, $items, $range, $discount, $weight, $multi, $memo);
+        return $this->updateOrCreate($rule_id, $name, $desc, $qualifies, $items, $range, $discount, $weight, $multi);
     }
 
     public function getRules($with_values = true)
@@ -54,7 +60,7 @@ class EloquentRuleRepository implements RuleRepositoryContract {
 
     public function syncItems($rule_id, $items)
     {
-        RuleItem::where('rule_id', $rule_id)->delete();
+        RuleItem::query()->where('rule_id', $rule_id)->delete();
 
         $item_data = [];
         foreach ($items['values'] as $value) {
@@ -68,13 +74,17 @@ class EloquentRuleRepository implements RuleRepositoryContract {
 
     public function syncQualifies($rule_id, $qualifies)
     {
+        if (empty($qualifies['values']) || !$qualifies['values']) {
+            return 0;
+        }
+
         RuleQualify::query()->where('rule_id', $rule_id)->delete();
 
         $qua_data = [];
         foreach ($qualifies['values'] as $value) {
             $qua_data[] = ['value' => $value, 'count' => $qualifies['quantity']];
         }
-        
+
         $rule = $this->getRule($rule_id);
         $rule->qualifies()->createMany($qua_data);
     }
@@ -83,7 +93,7 @@ class EloquentRuleRepository implements RuleRepositoryContract {
     {
         RuleItem::query()->where('rule_id', $rule_id)->delete();
         RuleQualify::query()->where('rule_id', $rule_id)->delete();
-        Rule::find($rule_id)->delete();
+        Rule::query()->find($rule_id)->delete();
         return 1;
     }
